+++ CnCTDRAMapEditor/SoleSurvivor/GameInfoSole.cs
@@ -65,7 +65,9 @@ namespace MobiusEditor.SoleSurvivor
         public override int MaxTerrain => Constants.MaxTerrainClassic;
         public override int MaxUnits => Constants.MaxUnitsClassic;
         public override int MaxTriggers => Constants.MaxTriggersClassic;
+        public override int MaxTriggerNameLength => Constants.MaxTriggerNameLength;
         public override int MaxTeams => Constants.MaxTeamsClassic;
+        public override int MaxTeamNameLength => Constants.MaxTeamNameLength;
         public override int MaxTeamClasses => Globals.MaxTeamClasses;
         public override int MaxTeamMissions => Globals.MaxTeamMissions;
         public override OverlayTypeFlag OverlayIconType => OverlayTypeFlag.FlagPlace;